Moved methods from the Modifier_Manager to an abstract class to make more things DRY.
