+++ src/Tickets/Order_Modifiers/Modifiers/Modifier_Manager.php
@@ -118,70 +118,4 @@ class Modifier_Manager {
 	public function render_edit_screen( array $context ): mixed {
 		return $this->strategy->render_edit( $context );
 	}
-
-	/**
-	 * Converts a decimal amount to its value in cents.
-	 *
-	 * This method is used to convert a floating-point amount (e.g., 23.00) into an integer representing cents.
-	 *
-	 * @since TBD
-	 *
-	 * @param float $amount The amount to convert.
-	 *
-	 * @return int The amount converted to cents.
-	 */
-	public function convert_to_cents( float $amount ): int {
-		return (int) round( floatval( $amount ) * 100 );
-	}
-
-	/**
-	 * Converts an amount in cents to a formatted decimal string.
-	 *
-	 * This method is used to convert an integer amount in cents (e.g., 2300) into a string with two decimal points (e.g., 23.00).
-	 *
-	 * @since TBD
-	 *
-	 * @param int $cents The amount in cents.
-	 *
-	 * @return string The formatted decimal string representing the amount.
-	 */
-	public function convert_from_cents( int $cents ): string {
-		return number_format( $cents / 100, 2, '.', '' );
-	}
-
-	/**
-	 * Generates a unique alphanumeric slug of 7 characters.
-	 *
-	 * The slug will be checked for uniqueness in the database before being returned.
-	 *
-	 * @since TBD
-	 *
-	 * @return string The unique slug.
-	 */
-	public function generate_unique_slug(): string {
-		$slug_length = 7;
-
-		// Generate a random alphanumeric slug.
-		do {
-			// Generate random bytes and convert them to an alphanumeric string.
-			$random_string = substr( base_convert( bin2hex( random_bytes( 4 ) ), 16, 36 ), 0, $slug_length );
-		} while ( ! $this->is_slug_unique( $random_string ) );
-
-		return $random_string;
-	}
-
-	/**
-	 * Checks whether a slug is unique in the database.
-	 *
-	 * @since TBD
-	 *
-	 * @param string $slug The slug to check for uniqueness.
-	 *
-	 * @return bool True if the slug is unique, false otherwise.
-	 */
-	protected function is_slug_unique( string $slug ): bool {
-		$existing_slug = $this->strategy->find_by_slug( $slug );
-
-		return null === $existing_slug;
-	}
 }
